+++ changelog.md
@@ -1,6 +1,6 @@
 # Fizz update changelog
 
-## Version 0.7.0
+## Version 1.0.0
 
 **Whats new:**